fix: localstorage undefined issue fixed

+++ src/utils/helper.ts
@@ -31,11 +31,17 @@ export const sendNotifications = (type: string, message: string) => {
 };
 
 export const storeInStorage = (key: string, value: string) => {
-  return localStorage.setItem(key, value);
+  if (typeof window !== "undefined") {
+    return localStorage.setItem(key, value);
+  }
 };
 export const getFromStorage = (key: string) => {
-  return localStorage.getItem(key);
+  if (typeof window !== "undefined") {
+    return localStorage.getItem(key);
+  }
 };
 export const clearStorage = () => {
-  return localStorage.clear();
+  if (typeof window !== "undefined") {
+    return localStorage.clear();
+  }
 };
